requirements.txt
azure-cognitiveservices-speech==1.13.0
azure-storage-blob==2.1.0
streamlit==0.67.1
scipy==1.5.1
pyyaml==5.3.1
ffmpeg-python==0.2.0
pydub==0.24.1